New_conc/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-openai==0.0.5
pandas==2.2.0
numpy==1.26.3
yfinance==0.2.35
python-dotenv==1.0.0
pydantic==2.5.3
httpx==0.26.0
